pyproject.toml
[project]
name = "tg-rag-smolagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ilya Lubenets", email = "xfilinxcl@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "docling>=2.28.0",
    "docling-parse>=4.0.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-docling>=0.2.0",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.11",
    "langchain-qdrant>=0.2.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=22.0",
    "smolagents>=1.12.0",
]

[project.scripts]
serve = "tg_rag_smolagents.bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
